[department-of-veterans-affairs/#orchid-107] Reorganize directory structure to be more domain driven
