+++ src/applications/ask-a-question/app-entry.jsx
@@ -1,10 +1,10 @@
 import 'platform/polyfills';
-import './sass/ask-a-question.scss';
+import './form/config/sass/ask-a-question.scss';
 
 import startApp from 'platform/startup';
 
-import routes from './routes';
-import reducer from './reducers';
+import routes from './form/config/routes';
+import reducer from './form/config/reducers';
 import manifest from './manifest.json';
 
 startApp({
