--- src/applications/ask-a-question/config/pages/index.js
@@ -1,5 +1,0 @@
-import contactInformationPage from './contactInformationPage';
-import inquiryPage from './inquiryPage';
-import veteranInformationPage from './veteranInformationPage';
-
-export { contactInformationPage, inquiryPage, veteranInformationPage };
