+++ src/applications/ask-a-question/form/config/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import RoutedSavableApp from 'platform/forms/save-in-progress/RoutedSavableApp';
-import formConfig from '../config/form';
+import formConfig from './form';
 
 export default function App({ location, children }) {
   return (
