+++ src/applications/ask-a-question/form/config/form.js
@@ -1,9 +1,9 @@
 // Example of an imported schema:
-import fullSchema from '../0873-schema.json';
+import fullSchema from '../../0873-schema.json';
 // In a real app this would be imported from `vets-json-schema`:
 // import fullSchema from 'vets-json-schema/dist/0873-schema.json';
-import IntroductionPage from '../containers/IntroductionPage';
-import ConfirmationPage from '../containers/ConfirmationPage';
+import IntroductionPage from '../introduction/IntroductionPage';
+import ConfirmationPage from '../confirmation/ConfirmationPage';
 import {
   contactInformationPage,
   inquiryPage,
@@ -24,7 +24,7 @@ import {
   submitButtonText,
   veteranInformationChapterTitle,
   veteranInformationPageTitle,
-} from '../content/labels';
+} from './labels';
 
 const {
   fullName,
