+++ src/applications/ask-a-question/form/config/pages/index.js
@@ -1,0 +1,5 @@
+import contactInformationPage from '../../contactInformation/contactInformationPage';
+import inquiryPage from '../../inquiry/inquiryPage';
+import veteranInformationPage from '../../veteran/veteranInformationPage';
+
+export { contactInformationPage, inquiryPage, veteranInformationPage };
