+++ src/applications/ask-a-question/form/config/reducers/index.js
@@ -1,4 +1,4 @@
-import formConfig from '../config/form';
+import formConfig from '../form';
 import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
 
 export default {
