+++ src/applications/ask-a-question/form/config/routes.jsx
@@ -1,6 +1,6 @@
 import { createRoutesWithSaveInProgress } from 'platform/forms/save-in-progress/helpers';
-import formConfig from './config/form';
-import App from './containers/App.jsx';
+import formConfig from './form';
+import App from './App.jsx';
 
 const route = {
   path: '/',
