+++ src/applications/ask-a-question/form/inquiry/topic/topic.js
@@ -1,8 +1,8 @@
 import _ from 'lodash/fp';
 import { createSelector } from 'reselect';
 
-import fullSchema from '../../0873-schema.json';
-import { topicTitle } from '../../content/labels';
+import fullSchema from '../../../0873-schema.json';
+import { topicTitle } from '../../config/labels';
 import {
   vaMedicalCentersLabels,
   vaMedicalCentersValues,
