+++ src/applications/ask-a-question/form/veteran/dependent/dependentInformationUI.js
@@ -1,5 +1,5 @@
 import { uiSchema } from 'platform/forms/definitions/address';
-import SectionHeader from '../../content/SectionHeader';
+import SectionHeader from '../../config/components/SectionHeader';
 import emailUI from 'platform/forms-system/src/js/definitions/email';
 import phoneUI from 'platform/forms-system/src/js/definitions/phone';
 import _ from 'lodash';
@@ -10,7 +10,7 @@ import {
   dependentsLastName,
   emailTitle,
   streetAddress,
-} from '../../content/labels';
+} from '../../config/labels';
 
 const formFields = {
   first: 'first',
