+++ src/applications/ask-a-question/form/veteran/status/veteranStatusUI.js
@@ -6,7 +6,7 @@ import {
   relationshipToVeteranTitle,
   veteranStatusSectionDescription,
   veteranStatusTitle,
-} from '../../content/labels';
+} from '../../config/labels';
 
 const formFields = {
   veteranStatus: 'veteranStatus',
